xorg: disable mtrack driver
it feels like this declaration was added because it enabled multitouch
capability to whatever touchpad driver which is enabled (as the name
suggests) at the moment but turns out that is not the case.  this
declaration will instead enable the mtrack touchpad driver and this
driver doesn't look maintained/supported anymore. this declaration
causes xorg to load the mtrack driver for the touchpad instead of
libinput and hence KDE is unable to configure the touchpad.

hence remove this declaration to disable because this driver already
defaults to being disabled.

refs:
https://github.com/NixOS/nixpkgs/blob/release-19.09/nixos/modules/services/x11/hardware/multitouch.nix which loads
https://github.com/NixOS/nixpkgs/blob/master/pkgs/os-specific/linux/xf86-input-mtrack/default.nix which points to
https://github.com/BlueDragonX/xf86-input-mtrack